Reorganize java extensions module
